// src/app/Constants/interface.tsx
export interface Interface {
    _id: string;
    name: string;
    description: any;
    image: string;
    caption: string;
    tag: String
    attribute: string;
    about: string;
    question: string;
    answer: any;
    university: string;
    degree: string;
    passing: any;
    company: string;
    role: string;
    working_period: any;
    link: any
}